daemon: fully rework getwork server
