+++ xelis_daemon/Cargo.toml
@@ -12,7 +12,6 @@ clap = { version = "4.5.2", features = ["derive"] }
 bytes = "1"
 actix = "0.13.0"
 actix-web = "4"
-actix-web-actors = "4"
 sled = "0.34.7"
 lru = "0.12.3"
 async-recursion = "1"
